Added user authentication to todos
Todos now only work with the user that created them

+++ server/server.js
@@ -17,9 +17,10 @@ const app = express();
 
 app.use(bodyParser.json());
 
-app.post('/todos', (req, res) => {
+app.post('/todos', authenticate, (req, res) => {
   const todo = new Todo({
     text: req.body.text,
+    _creator: req.user._id,
   });
   todo.save().then((doc) => {
     res.send(doc);
@@ -28,24 +29,24 @@ app.post('/todos', (req, res) => {
   });
 });
 
-app.get('/todos', (req, res) => {
-  Todo.find().then((todos) => {
+app.get('/todos', authenticate, (req, res) => {
+  Todo.find({ _creator: req.user._id }).then((todos) => {
     res.send({ todos });
   }, (e) => {
     res.status(400).send(e);
   })
 });
 
-app.get('/todos/:id', (req, res) => {
+app.get('/todos/:id', authenticate, (req, res) => {
   const id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
-    return res.status(404).send({ error: 'Invalid ID' });
+    return res.status(404).send({ error: 'Invalid Todo ID' });
   }
 
-  Todo.findById(id).then((todo) => {
+  Todo.findOne({ _id: id, _creator: req.user._id }).then((todo) => {
     if (!todo) {
-      return res.status(404).send({ error: 'ID not found' });
+      return res.status(404).send({ error: 'Todo not found' });
     }
 
     res.send({ todo });
@@ -54,16 +55,16 @@ app.get('/todos/:id', (req, res) => {
   });
 });
 
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id', authenticate, (req, res) => {
   const id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
-    return res.status(404).send({ error: 'Invalid ID' });
+    return res.status(404).send({ error: 'Invalid Todo ID' });
   }
 
-  Todo.findOneAndDelete({ _id: id }).then((todo) => {
+  Todo.findOneAndDelete({ _id: id, _creator: req.user._id }).then((todo) => {
     if (!todo) {
-      return res.status(404).send({ error: 'ID not found' });
+      return res.status(404).send({ error: 'Todo not found' });
     }
 
     res.send({ todo });
@@ -72,12 +73,12 @@ app.delete('/todos/:id', (req, res) => {
   });
 });
 
-app.patch('/todos/:id', (req, res) => {
+app.patch('/todos/:id', authenticate, (req, res) => {
   const id = req.params.id;
   const body = _.pick(req.body, ['text', 'completed']);
 
   if (!ObjectID.isValid(id)) {
-    return res.status(404).send({ error: 'Invalid ID' });
+    return res.status(404).send({ error: 'Invalid Todo ID' });
   }
 
   if (_.isBoolean(body.completed) && body.completed) {
@@ -87,13 +88,13 @@ app.patch('/todos/:id', (req, res) => {
     body.completedAt = null;
   }
 
-  Todo.findOneAndUpdate({ _id: id }, {
+  Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, {
     $set: body,
   }, {
     new: true,
   }).then((todo) => {
     if (!todo) {
-      return res.status(404).send({ error: 'ID not found' });
+      return res.status(404).send({ error: 'Todo not found' });
     }
 
     res.send({ todo });
